requirements.txt
aiogram
requests
beautifulsoup4
ddgs
tldextract
transformers
langchain-ollama
langchain-qdrant
pandas
numpy
pydantic
loguru
pyyaml
omegaconf
hydra-core
ipykernel
ipywidgets
ollama
langchain
langgraph
redis
langchain_community
asyncio
regex
apache-airflow
apscheduler
sentence-transformers
langchain-huggingface
undetected-chromedriver
duckduckgo-search
openai
icrawler
googlesearch-python
selenium
google-api-python-client
Google-Images-Search
qdrant-client
elasticsearch
fastapi
fastmcp
python-dotenv
